nrf52833_pac::radio
Module cteinlineconf

Configuration for CTE inline mode

Structs

CTEINLINECONF_SPEC

Configuration for CTE inline mode

R

Register CTEINLINECONF reader

W

Register CTEINLINECONF writer

Enums

CTEERRORHANDLING_A

Sampling/switching if CRC is not OK

CTEINFOINS1_A

CTEInfo is S1 byte or not

CTEINLINECTRLEN_A

Enable parsing of CTEInfo from received packet in BLE modes

CTEINLINERXMODE1US_A

Spacing between samples for the samples in the SWITCHING period when CTEINLINEMODE is set

CTEINLINERXMODE2US_A

Spacing between samples for the samples in the SWITCHING period when CTEINLINEMODE is set

CTETIMEVALIDRANGE_A

Max range of CTETime

Type Aliases

CTEERRORHANDLING_R

Field CTEERRORHANDLING reader - Sampling/switching if CRC is not OK

CTEERRORHANDLING_W

Field CTEERRORHANDLING writer - Sampling/switching if CRC is not OK

CTEINFOINS1_R

Field CTEINFOINS1 reader - CTEInfo is S1 byte or not

CTEINFOINS1_W

Field CTEINFOINS1 writer - CTEInfo is S1 byte or not

CTEINLINECTRLEN_R

Field CTEINLINECTRLEN reader - Enable parsing of CTEInfo from received packet in BLE modes

CTEINLINECTRLEN_W

Field CTEINLINECTRLEN writer - Enable parsing of CTEInfo from received packet in BLE modes

CTEINLINERXMODE1US_R

Field CTEINLINERXMODE1US reader - Spacing between samples for the samples in the SWITCHING period when CTEINLINEMODE is set

CTEINLINERXMODE1US_W

Field CTEINLINERXMODE1US writer - Spacing between samples for the samples in the SWITCHING period when CTEINLINEMODE is set

CTEINLINERXMODE2US_R

Field CTEINLINERXMODE2US reader - Spacing between samples for the samples in the SWITCHING period when CTEINLINEMODE is set

CTEINLINERXMODE2US_W

Field CTEINLINERXMODE2US writer - Spacing between samples for the samples in the SWITCHING period when CTEINLINEMODE is set

CTETIMEVALIDRANGE_R

Field CTETIMEVALIDRANGE reader - Max range of CTETime

CTETIMEVALIDRANGE_W

Field CTETIMEVALIDRANGE writer - Max range of CTETime

S0CONF_R

Field S0CONF reader - S0 bit pattern to match

S0CONF_W

Field S0CONF writer - S0 bit pattern to match

S0MASK_R

Field S0MASK reader - S0 bit mask to set which bit to match

S0MASK_W

Field S0MASK writer - S0 bit mask to set which bit to match