nrf52833_pac::radio::cteinlineconf
Struct R

pub struct R(/* private fields */);

Register CTEINLINECONF reader

Implementations

impl R

pub fn cteinlinectrlen(&self) -> CTEINLINECTRLEN_R

Bit 0 - Enable parsing of CTEInfo from received packet in BLE modes

pub fn cteinfoins1(&self) -> CTEINFOINS1_R

Bit 3 - CTEInfo is S1 byte or not

pub fn cteerrorhandling(&self) -> CTEERRORHANDLING_R

Bit 4 - Sampling/switching if CRC is not OK

pub fn ctetimevalidrange(&self) -> CTETIMEVALIDRANGE_R

Bits 6:7 - Max range of CTETime

pub fn cteinlinerxmode1us(&self) -> CTEINLINERXMODE1US_R

Bits 10:12 - Spacing between samples for the samples in the SWITCHING period when CTEINLINEMODE is set

pub fn cteinlinerxmode2us(&self) -> CTEINLINERXMODE2US_R

Bits 13:15 - Spacing between samples for the samples in the SWITCHING period when CTEINLINEMODE is set

pub fn s0conf(&self) -> S0CONF_R

Bits 16:23 - S0 bit pattern to match

pub fn s0mask(&self) -> S0MASK_R

Bits 24:31 - S0 bit mask to set which bit to match