nrf52833_pac::radio::dfectrl1
Type Alias TSAMPLESPACINGREF_R

pub type TSAMPLESPACINGREF_R = FieldReader<u8, TSAMPLESPACINGREF_A>;

Field TSAMPLESPACINGREF reader - Interval between samples in the REFERENCE period

Aliased Type

struct TSAMPLESPACINGREF_R { /* private fields */ }

Implementations

impl TSAMPLESPACINGREF_R

pub fn variant(&self) -> Option<TSAMPLESPACINGREF_A>

Get enumerated values variant

pub fn is_4us(&self) -> bool

Checks if the value of the field is _4US

pub fn is_2us(&self) -> bool

Checks if the value of the field is _2US

pub fn is_1us(&self) -> bool

Checks if the value of the field is _1US

pub fn is_500ns(&self) -> bool

Checks if the value of the field is _500NS

pub fn is_250ns(&self) -> bool

Checks if the value of the field is _250NS

pub fn is_125ns(&self) -> bool

Checks if the value of the field is _125NS